Doc update for Resize to correct some wrong information (#2948)
* Doc update for Resize.
The scales is set as Single input, it's not possible to set an empty string name for it because we have schema check for Single input. The doc is not correct. The only way is to set a nonempty string as the name but set it to empty data (empty shape).
Another option is to set scales as Optional input, but it requires opset version change.

* use gen_doc.py to generate the docs